Fix a warning
```
../gc.c:2342:45: warning: comparison of integers of different signs: 'short' and 'size_t' (aka 'unsigned long') [-Wsign-compare]
    GC_ASSERT(size_pools[pool_id].slot_size == slot_size);
    ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~^~~~~~~~~~~~~
```

Add cast to short, because `GC_ASSERT`s in `size_pool_for_size`
already use cast to short.

+++ gc.c
@@ -2339,7 +2339,7 @@ size_pool_slot_size(unsigned char pool_id)
 
 #if RGENGC_CHECK_MODE
     rb_objspace_t *objspace = &rb_objspace;
-    GC_ASSERT(size_pools[pool_id].slot_size == slot_size);
+    GC_ASSERT(size_pools[pool_id].slot_size == (short)slot_size);
 #endif
 
     return slot_size;
